モーダル部分記述整理
